Check in shared travis scripts
Take `pulumi/pulumi`'s scripts that are used in CI and share them
here. We'll move to a model where travis clones this repository and
then invokes these scripts directly, instead of having each repository
have to clone a bunch of logic around how to install toolchains, what
versions of things to install and the like.