Open source `hack-vendor`
`hack-vendor` is a small script that makes it quick to override a
vendored dependency (managed by `dep`) with a local version. It works
by spinning up a local web server that exposes $GOPATH/src and then
modifies the `Gopkg.toml` file of a project to override the source of
the package you want to modify to point at this local server and does
a restore.

The web-server listens on port 80, due to limitiations with `dep` when
this script was authored. It appears that newer versions would allow
sources to actually have a non standard port, so in the future we may
be able to modify the scripts to just have the web-server listen on
some non standard port, which would prevent the need to sudo.