Add a script for testing "at HEAD"
The intention of this new script is to allow us to run our node tests
"at HEAD". When run, this script does a few things:

1. For every `@pulumi` package, figure out the latest version, by
   looking at what the `dev` tag points at.

2. Install the latest dev build of the CLI (i.e. the one that matches
   the "dev" tag of @pulumi/pulumi).

3. Set an environment variaible that has information about the latest
   versions of all the packages.

4. Call `make travis_${TRAVIS_EVENT_TYPE}` like a normal Travis build
   would.

The integration tests in each repository will use the environment
variable that we've set to pass information about the latest versions
to the integration test framework, to ensure we use the newest
versions.