Adopt new node installer scripts
This change extracts the shared logic we have duplicated in every
terraform based provider repository into a common location.

In addition, it augments the publishing scripts to support
pulumi/pulumi#2155.

Previously, the poorly named `promote.js` would rewrite the
`package.json` of a package before we were going to publish it. It
would add an `install` script to the `scripts` section of package.json
which called `pulumi plugin install resource <name> <version>` so that
when packages were restored, the plugin would be installed.

However, when installing on a machine where `pulumi` was not present,
this script would fail and cause the entire `npm install` or `yarn
install` to fail because `pulumi` could not be run.

Now, instead of invoking pulumi directly, we invoke a script which
ships with the package (and who's source lives here and is injected
into the package as we are preparing it to publish) that tries to
shell out to `pulumi` and handles the errors in a nicer manner.

This allows plugin install to work even if `pulumi` is not on the
machine or not on the path.