Add policy checks for packages before publishing
We want to ensure that when we publish a package, none of the
dependencies are pre-release dependencies (which we, for better or
worse, end up using often during cross cutting development across our
many repositories).

This change adds a new script, `check-package-dependencies` that will
run validation logic against the package.json we are about to publish
and use its exit code to communicate if the package should be
published or not.

Our shared publishing script uses this a pre-check before actually
doing the publish.