Calculate and set OS from outside of the subshell
install-common-toolchain.sh is defining and setting OS from within a
subshell while also referencing that variable in the parent shell. The
writes to the OS variable from within the subshell are not visible in
the parent shell and, as a result, we were not putting Python's bin
directory on the PATH on OSX.

This fixes the issue by defining OS in the parent shell and exporting
it, so that it's visible in the subshell and in the parent shell.