Add gomod-override
gomod-override is analagous to govendor-override, except it works on go.mod
files used with Go 1.11 modules instead of govendor.json files.

For now it has the following restrictions:

- Only the first constraint in the template with `gomod-override = true` will be
  processed
- Replace directives in the go.mod file are not honoured

This is sufficient to build a solvable set of dependency constraints for the
Google Cloud Beta provider, which was the initial aim. It's anticipated that
further work will be required over time.