Rework gomod-override
Rather than parsing the go.mod file in a target repository, we can instead use
the go tool to do this including all indirect dependencies.

Additionally, rather than copying internal packages of the Go tool, we can use
github.com/rogpeppe/go-internal instead.

This fixes gomod-override to work for repositories where not all of the
transitive dependencies are declared in the `go.mod` file of the target, which
were failing with the old approach.