Use unmodified version for non-Git repositories
Serveral of our upstreams now have a dependency on a Bazaar repository. Since
it's pinned to a specific revision, this looks like a short Git SHA which we try
to resolve using `go get` before finding the complete SHA. If we can't open the
repository resulting from `go get`, we should use the SHA component as is, since
it likely refers to another VCS revision identifier.