Handle replaced modules using modules.txt
This approach for handling replaced module is required less user input, and is
more correct than the (reverted) approach taken previously - instead of
specifying source and destination organizations, we read the entire replacement
from the modules.txt file and use that where appropriate.

This is necessary for when we replace import paths with forks of providers (e.g.
AWS, Azure).