`run-at-head` improvements
1. Add `--no-latest-cli` and `--no-latest-packages` as way to control
if the latest CLI or Packages are used during the run. This lets us
mix and match things like "current set of packages with dev CLI" and
"current released CLI with latest dev packages"

2. Remove Python overrides stuff. Sean added this code in the past but
nothing actually uses it. The list of packages is no longer complete
and rather than bringing it up to date, I'm just removing it for now.

3. Expand the list of nodejs packages we compute the latest versions
for. I computed this list by looking at the `@pulumi` scope on NPM,
removing any private packages.