Add a script for building package docs. (#82)
These changes add a new script that clones
https://github.com/pulumi/docs, regenerates the appropriate parts of the
docs site, and opens a pull request with the changes. Documentation will
only be regenerated for commits that have an appropriate tag. This
script should be run after packages have been published, and should be
run as part of the same job that performed the build.