Don't fail when a package has already been published
If the package we are trying to publish already exists, just continue
along our merry way. This happens for us in cases where we have a push
build for master racing with the tag build (when we tag HEAD of master
during a release) and can also happen if we need to restart a build.

NPM doesn't make it super easy to do this, so we throw a little bash
elbow grease at the problem. You might think that instead of looking
at the output of `npm info` we could instead look at the return code,
but sadly for us `npm info foo@some-version-that-doesnt-exist` does
not have a non zero exit code.